Refactor: split OpenAI and Telegram logic
